Move TableMenuControls into `controls` directory
